[ember] @ember/array as source of types, ember as re-export
