+++ types/ember__array/tsconfig.json
@@ -28,6 +28,9 @@
         "index.d.ts",
         "mutable.d.ts",
         "proxy.d.ts",
+        "-private/native-array.d.ts",
+        "-private/enumerable.d.ts",
+        "-private/mutable-enumerable.d.ts",
         "types/prototype-extensions.d.ts",
         "test/lib/assert.ts",
         "test/array.ts",
