+++ types/ember__array/types/prototype-extensions.d.ts
@@ -1,4 +1,6 @@
-import Ember from 'ember';
+// import Ember from 'ember';
 import Observable from '@ember/object/observable';
+import MutableArray from '@ember/array/mutable';
+import Copyable from '@ember/object/-private/copyable';
 
-export interface ArrayPrototypeExtensions<T> extends Ember.MutableArray<T>, Observable, Ember.Copyable {}
+export default interface ArrayPrototypeExtensions<T> extends MutableArray<T>, Observable, Copyable {}
